typescriptize PluginSetting
